// src/http/schemas/auth.ts
import { z } from 'zod';

export const registerSchema = z.object({
  name: z.string().min(3).max(100),
  email: z.string().email(),
  password: z.string().min(6),
  role: z.enum(['teacher', 'student']),
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string(),
});